refactor: test도 이름 변경
